vali: fitting drivers to the new irq method

+++ librt/libc/include/internal/_evt.h
@@ -6,7 +6,7 @@
 struct evt {
     unsigned int flags;
     unsigned int options;
-    size_t       initialValue;
+    unsigned int initialValue;
     atomic_int*  sync_address;
 };
